Use Dialog prompt to init the SketchEngine
When the player answer no, the widget close. If he opens it back, the
dialog will still show the same question, to allow him to start a project
if he wants.
The Main module (in view) receive the SketchEngine instead of the whole
package. This is to make sure that only stuff in SketchEngine get called
by the interface.