Connected sections to sheets added
Fix os.time in LuaMetadata. Clean SketchEngine from creating a
random node. SectionView is not more versatile with its AddElement
method (using GuiObject instead of a preset TextButton). Adds
footer method (AddFooter) to SectionView. GameEditor now has input
to create new Sheets as a footer to the section. Implements a
SectionButton in Templates. Adds a preference button in toolbar
in Main and connects with sheets. Adds method to get the engine
from main.